fix: unify legacy render pipeline through container pipeline
