+++ packages/editor/src/editor/render-legacy-callbacks-context.ts
@@ -1,0 +1,10 @@
+import {createContext} from 'react'
+import type {
+  RenderChildFunction,
+  RenderPlaceholderFunction,
+} from '../types/editor'
+
+export const RenderLegacyCallbacksContext = createContext<{
+  renderChild?: RenderChildFunction
+  renderPlaceholder?: RenderPlaceholderFunction
+}>({})
