+++ packages/editor/src/editor/render.element.tsx
@@ -6,10 +6,8 @@ import {isTextBlock} from '@portabletext/schema'
 import {useSelector} from '@xstate/react'
 import {useContext, type ReactElement} from 'react'
 import type {DropPosition} from '../behaviors/behavior.core.drop-position'
-import {isInline} from '../node-traversal/is-inline'
 import type {Path} from '../slate/interfaces/path'
 import type {RenderElementProps} from '../slate/react/components/editable'
-import {useSlateStatic} from '../slate/react/hooks/use-slate-static'
 import type {
   RenderBlockFunction,
   RenderChildFunction,
@@ -19,11 +17,8 @@ import type {
 import {ContainerScopeContext} from './container-scope-context'
 import {EditorActorContext} from './editor-actor-context'
 import type {EditorSchema} from './editor-schema'
-import {RenderBlockObject} from './render.block-object'
 import {RenderContainer} from './render.container'
 import {RenderContainerChild} from './render.container-child'
-import {RenderInlineObject} from './render.inline-object'
-import {RenderTextBlock} from './render.text-block'
 import {buildScopedName, findByScope} from './scoped-config-lookup'
 
 export function RenderElement(props: {
@@ -42,21 +37,12 @@ export function RenderElement(props: {
 }) {
   const editorActor = useContext(EditorActorContext)
   const containerScope = useContext(ContainerScopeContext)
-  const slateStatic = useSlateStatic()
   const schema = props.schema
 
   const scopedTypeName = buildScopedName(containerScope, props.element._type)
-  const isInlineObject = isInline(slateStatic, props.path)
-  const leafScopedName = isInlineObject
-    ? buildScopedName(containerScope, `block.${props.element._type}`)
-    : scopedTypeName
 
   const containerConfig = useSelector(editorActor, (s) =>
     findByScope(s.context.containerConfigs, scopedTypeName),
-  )
-
-  const leafConfig = useSelector(editorActor, (s) =>
-    findByScope(s.context.leafConfigs, leafScopedName),
   )
 
   if (containerConfig) {
@@ -86,71 +72,43 @@ export function RenderElement(props: {
 
   if (isTextBlock({schema}, props.element)) {
     return (
-      <RenderTextBlock
-        attributes={props.attributes}
-        dropPosition={
-          props.dropPosition?.blockKey === props.element._key
-            ? props.dropPosition.positionBlock
-            : undefined
-        }
-        element={props.element}
-        path={props.path}
-        readOnly={props.readOnly}
-        renderBlock={props.renderBlock}
-        renderListItem={props.renderListItem}
-        renderStyle={props.renderStyle}
-        schema={schema}
-        spellCheck={props.spellCheck}
-        textBlock={props.element}
-      >
-        {props.children}
-      </RenderTextBlock>
-    )
-  }
-
-  if (leafConfig) {
-    return (
-      <RenderContainerChild
+      <RenderContainer
         attributes={props.attributes}
         element={props.element}
+        legacyCallbacks={{
+          dropPosition:
+            props.dropPosition?.blockKey === props.element._key
+              ? props.dropPosition.positionBlock
+              : undefined,
+          renderBlock: props.renderBlock,
+          renderListItem: props.renderListItem,
+          renderStyle: props.renderStyle,
+          schema,
+          spellCheck: props.spellCheck,
+        }}
         path={props.path}
       >
         {props.children}
-      </RenderContainerChild>
-    )
-  }
-
-  if (isInlineObject) {
-    return (
-      <RenderInlineObject
-        attributes={props.attributes}
-        element={props.element}
-        path={props.path}
-        readOnly={props.readOnly}
-        renderChild={props.renderChild}
-        schema={schema}
-      >
-        {props.children}
-      </RenderInlineObject>
+      </RenderContainer>
     )
   }
 
   return (
-    <RenderBlockObject
+    <RenderContainerChild
       attributes={props.attributes}
-      blockObject={props.element}
-      dropPosition={
-        props.dropPosition?.blockKey === props.element._key
-          ? props.dropPosition.positionBlock
-          : undefined
-      }
       element={props.element}
+      legacyCallbacks={{
+        dropPosition:
+          props.dropPosition?.blockKey === props.element._key
+            ? props.dropPosition.positionBlock
+            : undefined,
+        readOnly: props.readOnly,
+        renderBlock: props.renderBlock,
+        renderChild: props.renderChild,
+      }}
       path={props.path}
-      readOnly={props.readOnly}
-      renderBlock={props.renderBlock}
-      schema={schema}
     >
       {props.children}
-    </RenderBlockObject>
+    </RenderContainerChild>
   )
 }
