--- packages/editor/src/editor/render.text.tsx
@@ -1,18 +1,0 @@
-import type {Editable} from '../slate/react/components/editable'
-
-export type RenderTextProps = Parameters<
-  NonNullable<React.ComponentProps<typeof Editable>['renderText']>
->[0]
-
-export function RenderText(props: RenderTextProps) {
-  return (
-    <span
-      {...props.attributes}
-      data-child-key={props.text._key}
-      data-child-name={props.text._type}
-      data-child-type="span"
-    >
-      {props.children}
-    </span>
-  )
-}
